fix: Erro ao deletar vendedor com vendas associadas a ele.
Método delete agora lança um body na response, informando que as vendas associadas ao vendedor devem ser transferidas ou apagadas para conseguir deletar o vendedor.

+++ src/main/java/com/orgvendas/domain/service/VendedorService.java
@@ -72,8 +72,13 @@ public class VendedorService {
 
         if (vendedorOptional.isPresent()) {
             Vendedor vendedor = vendedorOptional.get();
-            vendedorRepository.deleteById(id);
-            return ResponseEntity.ok().body("O(a) vendedor(a) " + vendedor.getNome()+ " foi deletado(a). ");
+            if(vendedor.getTotalDeVendas()>0){
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("Para deletar esse(a) vendedor(a) é necessário que as vendas associadas a ele(a) sejam tranferidas ou deletadas");
+            }else{
+                vendedorRepository.deleteById(id);
+                return ResponseEntity.ok().body("O(a) vendedor(a) " + vendedor.getNome()+ " foi deletado(a). ");
+            }
+
         } else{
             return ResponseEntity.notFound().build();
         }
